File::copy () should return false on I/O errors
File::copy would return true even if it was unable to actually copy the
data from the source file to the destination file.

This function also does not need a separate access check since ifstream will
already set the fail bit if something went wrong with the open.

GothenburgBitFactory/timewarrior#155