Support capitalized day and month names
As pointed out in [1] it used to be possible to use capitalized day and
month names in Taskwarrior. Since switching to the `Datetime` format in
[2], this is no longer possible.
This commit tries to fix that by ensuring the `Pig::skipPartial` method
always compares lower-case characters (since this is what `reference`
always contains).

Unittests are added to assert this behavior.

[1]: https://github.com/GothenburgBitFactory/taskwarrior/issues/2160
[2]: https://github.com/GothenburgBitFactory/taskwarrior/commit/434e95f560b025f07d2fa5fb9a3733fa1ddb4d42