format: Make sure to never initialize strings with negative fill length
The std::string (size_t n, char c) constructor uses size_t type, which
is a unsigned integer type. If a negative signed integer is passed as
first argument, the number underflows during the conversion and we end
up generating a very long string filled with c, which usually leads to
OOMKilled event.

Closes: GothenburgBitFactory/taskwarrior#2389.