Call fsync/fdatasync on File::close.
Improves durability in the event of sudden crash - fclose() does not
guarantee that data is actually flushed to disk. This is first step in
durability improvement plan discussed in
https://github.com/GothenburgBitFactory/taskwarrior/issues/152.

The main user - TaskWarriror's TDB2 sequentlally calls close() at the end of
each command on four files. If the files were not modified, fsync() should be
close to no-op - we'll pay small price to syscall/context switch.

However if the files are dirty we will pay some performance penalty.
Most of the time fsync() will wait for 2 blocks to be persisted while
fdatasync() will usually require just one. On HDD this could add about
10ms per block, so on 4 dirty files in non-Linux system this could
add up to 80ms of latency to each modifying command. On Linux on SSD
this could be under 4ms (assuming SSD is in good state and does not trigger gc
on every write).

This seems small enough price to pay.