Implement empty AtomicFile.
It currently defines calls necessary for compiling with TDB2 and simply
proxies them to regular File.

Some naive tests also added to make sure basics work.

As part of this change Directory changed to be child of Path. Most of
the File methods did not made sense for it. It compiles with TaskWarrior
but other users may see missing methods. Most of shared methods should
be moved to Path.