Replace dayOfWeek determination with a canonical version from RFC3339
Whilst working on a week parsing issue, while reviewing the Zeller's
Congruence implementation that determines day of week, some questions
came up, such as:

 - in all the literature I could find, the algorithms were using a term
   like "(13 * (m + 1)) / 5", not "(13 * m - 1) / 5".  This made me
   confused (see for example
   https://en.wikipedia.org/wiki/Zeller%27s_congruence)

 - the return for weekday number in the algorithm uses 0 = Saturday,
   which does not appear to be corrected for (libshared code presumes
   0 = Sunday).

Rather than figure out the correctness myself, this patch replaces the
algorithm with one specified in RFC3339, which should be canonical, and
which is 0-Sunday based.

It passed all the same tests.  Probably more tests should be added.