Test date string parsing with both weekstart 0 and weekstart 1
We make a loop over all the testParse() calls, running them twice: first
with weekstart 0, then with weekstart 1.

We have to offset by one day the time_t variables given for comparison
(last parameter to testParse()), corresponding to the first day in week,
for weeks given without a day offset.  So local1, utc1, f_local1, f_utc1
are tweaked by one day for the second loop (weekstart 1).

Likewise, absolute dates like 2013-12, which test against these offset
values, need to be de-offset depending on which loop iteration so there
is an inline adjustment to the time_t value passed for those specific 4
test calls (2013-12 and 9850-12, with and without the '-').

Patch is best viewed with "diff -w".