Fix wrong week number output when weekstart=0
It's unclear why this week number increment was added, but it makes the
week number wrong when weekstart=0.  The condition was never true before
because weekstart was always 1, as initialized at the top of the file,
and Taskwarrior was not passing rc.weekstart in, so the [supposed] bug
never manifested.  We are now working on propagating rc.weekstart to
Datetime::weekstart, and sure enough, when weekstart=0, the week number
is ahead by one.