Fix timegm() to use HAVE_TIMEGM from cmake so it uses libc version
There is a preprocessor define to use the system version of timegm()
rather than a [slow] fallback function implemented in libshared.  The
function Datetime::timegm() conditionally manifests if said define is
unset, but the code won't actually consult the value (provided by
cmake), unless we include the cmake header.

(This function is only used by Timewarrior, so Taskwarrior need not peg
to a new libshared just for this fix.)